code cleanup and some fixes

+++ handlers.go
@@ -4,7 +4,7 @@ type (
 	ErrorHandlerFunc       func(err error)
 	SpawnPartHandlerFunc   func(hash string, ioff, foff int64)
 	RespawnPartHandlerFunc func(hash string, ioffNew, foffNew int64)
-	ProgressHandlerFunc    func(hash string, nread int64)
+	ProgressHandlerFunc    func(hash string, nread int)
 	OnCompleteHandlerFunc  func(hash string, tread int64)
 )
